requirements.txt
# Atlas-GRAG Dependencies
# Graph Retrieval Augmented Generation for Supply Chain Risk Analysis

# LLM & Embeddings (Local)
ollama>=0.3.0

# Graph Database
neo4j>=5.0.0

# Vector Database
chromadb>=0.4.0

# LangChain Framework
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0

# Web Framework
streamlit>=1.30.0

# Environment Configuration
python-dotenv>=1.0.0

# Async Support
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# Type Hints & Validation
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0

# Development
black>=24.0.0
mypy>=1.0.0
ruff>=0.4.0
